+++ app/src/main/java/com/obsidian/plugins/task_notifier/core/bo/PersistentObsidianReminderBO.kt
@@ -1,0 +1,6 @@
+package com.obsidian.plugins.task_notifier.core.bo
+
+import com.obsidian.plugins.task_notifier.core.bo.ObsidianReminderBO
+
+class PersistentObsidianReminderBO(val reminders: List<ObsidianReminderBO>) {
+}
