+++ app/src/main/java/com/obsidian/plugins/task_notifier/plugin/ObsidianPluginManager.kt
@@ -4,6 +4,7 @@ import android.net.Uri
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonDeserializer
+import com.obsidian.plugins.task_notifier.core.bo.ObsidianReminderBO
 import com.obsidian.plugins.task_notifier.plugin.dto.v1.ObsidianReminderPluginConfigDTO
 import com.obsidian.plugins.task_notifier.plugin.dto.v1.ObsidianReminderPluginDate
 import com.obsidian.plugins.task_notifier.utils.Logger
@@ -27,7 +28,6 @@ class ObsidianPluginManager {
 
     @JvmStatic
     fun processFile(json: String): ArrayList<ObsidianReminderBO> {
-      // get configs
       val jsonObject = Gson().fromJson(json, com.google.gson.JsonObject::class.java)
       val remindersObject = jsonObject.getAsJsonObject("settings")
       val dateTimeFormat = "YYYY-MM-DD HH:mm"
