--- app/src/main/java/com/obsidian/plugins/task_notifier/plugin/PersistentObsidianReminderBO.kt
@@ -1,4 +1,0 @@
-package com.obsidian.plugins.task_notifier.plugin
-
-class PersistentObsidianReminderBO(val reminders: List<ObsidianReminderBO>) {
-}
